export splitTitleAndContent function && apply store value
