+++ client/src/utils.ts
@@ -57,3 +57,19 @@ export const url = `${process.env.URL || 'http://localhost'}:${process.env.PORT 
 export const ORDER_WEIGHT = 100000;
 export const POSITION = { UP: -1, DOWN: 1, TOP: -2, BOTTOM: 2, INIT: 0 };
 export const calcMedium = (a: number, b: number): number => (a + b) / 2;
+export const splitTitleContent = (raw: string) => {
+	let title, content, tmp;
+
+	tmp = raw.split('<br/>');
+	if (tmp.length <= 1) {
+		title = tmp[0];
+		content = '';
+	} else {
+		title = tmp[0];
+		tmp.shift();
+		content = tmp.reduce((prev, now) => (prev += now + '<br/>'), '');
+		if (content.length >= 5) content = content.substring(0, content.length - 5);
+	}
+
+	return { title, content };
+};
